+++ src/input.h
@@ -1,6 +1,8 @@
 #ifndef INPUT_H
 #define INPUT_H
 
-char *read_bf_code(const char *filename, int *comma_counter);
+#include <stddef.h>
+
+char *read_bf_code(const char *filename, size_t *comma_counter);
 
 #endif
